backend/pyproject.toml
[project]
name = "powerbi_agents"
version = "0.1.0"
dependencies = [
    "google-adk",
    "google-generativeai",
    "google-cloud-bigquery",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "SQLAlchemy>=2.0",
    "psycopg[binary]",
    "alembic"
]
